backend/app/requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
appdirs==1.4.3
async-exit-stack==1.0.1
async-generator==1.10
asyncio-redis==0.15.1
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
broadcaster==0.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.54.1
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
motor==2.1.0
nest-asyncio==1.3.2
orjson==2.6.5
packaging==20.3
passlib==1.7.2
pathspec==0.8.0
pluggy==0.13.1
promise==2.3
py==1.8.1
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.1
pytest-asyncio==0.10.0
python-dotenv==0.12.0
python-multipart==0.0.5
python-slugify==4.0.0
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
Rx==1.6.1
six==1.14.0
starlette==0.13.2
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
typer==0.1.1
ujson==2.0.3
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
wcwidth==0.1.9
websocket-client==0.57.0
websockets==8.1
zipp==3.1.0
